Added a new private csr (m_bonfire) with currently only one bit
The bit is named sstep and is bit 0 of the register. When set the
processor will only execute one instruction after the next eret and then
emmit a breakpoint trap. It is for hard-ware assistet single step.
If an interrupt occurs immediatly after eret it has priority and is
executed without trapping.