reset fix, whitespace cleanup
Fixes behaviour at reset: The LLI Bus was enabled during reset and was
fetching the first instruction at the reset address all the time during
reset was asserted. On deasseration of reset the result of this fetch
was used instead of making a clean refetch.
The fix ensures that the bus remains idle during reset assertion and the
first instruction is fetched after reset deassertion